Swap sound buffers VBlank interrupt to stop playing garbage.
The screen transition makes the game loop miss a few vblanks while it
loads sprites into OAM. Swapping the audio bufferes during the screen
loop therefore also misses some VBlanks, which causes the DMA to
overrun the sound buffer resulting in garbage being played back for a
few frames.

Swapping the sound buffers in the interrupt means we never miss a
VBlank. We now zero the back buffer in the swap, so if we miss filling
the buffer, it'll playback silence instead of previously played data.

Also need to fill the back buffer in a critical_section so we don't
swap buffers while we're writing to the back buffer.